Move build_resources functionality into manager.py
